Add files via upload
Nowadays covid-19 is the talk of the town. Our project is based on a dataset named “COVID Dataset”. We
work on different types of models like Linear Regression. We analyze Lab Test, Confirmed Case and Death
Case of total 626 days (2020-April-04 to 2021-December-20) data. As our Intermediate Progress we use
different types of Exploratory Data Analysis (EDA) tools for analyzing these data such as the growth and
decay of covid-19 etc.